+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add `ArraySubset::to_ranges()`
+
 ## [0.17.0-beta.3] - 2024-09-26
 
 ### Highlights / Major Changes